Support more Unicode gems; Ruby 1.8 compatibility.
The Unicode upper and lowercase conversions can be provided by the
'unicode' gem or the native Java support in JRuby.  Support both
of those.  Also, the method names #id and #type are used throughout
the citeproc-ruby code.  Those are methods defined on Object in
Ruby 1.8.  Undefine them (if you need them, use #object_id and